folder_file: change type of new from short to int
As requested by Rocco Rutte in bug #2421 increase the type to int.

+++ browser.h
@@ -29,7 +29,7 @@ struct folder_file
   char *name;
   char *desc;
 
-  unsigned short new;
+  unsigned int new;
 #ifdef USE_IMAP
   char delim;
